added levels
(wrote a quick one-line fix for exploit that allowed infinite float mode by releasing space before timer ran out and then immediately pressing it again)

I couldn't figure out how to do a few things and fix a few bugs, so I'm writing them down here since most are relatively simple changes that should be easy for you since you're already familiar with the project.

To-do (if any of these are too hard and you feel it won't make enough of a difference, feel free to skip them):

- Decrease float timer to like 30-50% of what it is now

- Add a lose condition for only when not possessed (for level 1 spikes)

- Make loss condition only when not floating (immunity to spikes while floating)

- Allow de-possessing while midair

- Retry button / stage selector

- Ghost should pop out to the direction the player last moved

- For level 3: bounce mechanic for possessed basketball, add a win condition only when possessing


Level descriptions:

Level 1: Tile7 is supposed to be spikes. This possess square can move on spikes with no consequences, but the ghost dies if it touches it (unless in float mode). Player is supposed to possess the square, use it to get closer to the exit, and then exit possession, stand on square, and then float through wall into the win zone.

Level 2: Ghost should not fit through the gap between any two red tiles, but coin should. Player should go through the green portion in the center of the screen, then left, possess coin, go through the small gaps to return to beginning area, then go through the small gap to get to the win zone.

Level 3: Ghost should float up, possess basketball, then bounce off the platform at the bottom, and then float to the hoop. Should not have enough float to simply float from the first platform to the hoop; must bounce first. Bounce should be high--like 80-90% of the height it fell from, but not enough to make it to the hoop without some help from float mode.